Fix some layer id enum colliding warnings
C++20 under MSVC does not like seeing enum values between two enums getting operated on.
static cast to int fixes it for what we are trying to do anyway

+++ gerbview/gerbview_printout.cpp
@@ -85,7 +85,7 @@ void GERBVIEW_PRINTOUT::setupViewLayers( KIGFX::VIEW& aView, const LSET& aLayerS
     BOARD_PRINTOUT::setupViewLayers( aView, aLayerSet );
 
     for( LSEQ layerSeq = m_settings.m_LayerSet.Seq(); layerSeq; ++layerSeq )
-        aView.SetLayerVisible( GERBVIEW_LAYER_ID_START + *layerSeq, true );
+        aView.SetLayerVisible( static_cast<int>( GERBVIEW_LAYER_ID_START ) + *layerSeq, true );
 }
